dpsu api fixes
